Refactor the directory.
