+++ module_api_gateway/src/main/java/com/seonghun/apigateway/ApiGatewayApplication.java
@@ -1,17 +1,15 @@
-package org.example.prepurchase;
+package com.seonghun.apigateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
 
-@EnableFeignClients
 @SpringBootApplication
 @EnableDiscoveryClient
-public class ModuleUserServiceApplication {
+public class ApiGatewayApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ModuleUserServiceApplication.class, args);
+        SpringApplication.run(ApiGatewayApplication.class, args);
     }
     
 }
