+++ module_user_service/src/main/java/com/seonghun/module_user_service/ModuleUserServiceApplication.java
@@ -1,15 +1,17 @@
-package org.example.apigateway;
+package com.seonghun.module_user_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
+@EnableFeignClients
 @SpringBootApplication
 @EnableDiscoveryClient
-public class ApiGatewayApplication {
+public class ModuleUserServiceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ApiGatewayApplication.class, args);
+        SpringApplication.run(ModuleUserServiceApplication.class, args);
     }
     
 }
