+++ module_user_service/src/main/java/com/seonghun/module_user_service/domain/user/application/UserService.java
@@ -1,13 +1,12 @@
-package org.example.prepurchase.domain.user.application;
+package com.seonghun.module_user_service.domain.user.application;
 
-import jakarta.servlet.http.HttpServletResponse;
-import org.example.prepurchase.domain.user.domain.Users;
-import org.example.prepurchase.domain.user.dao.UserRepository;
-import org.example.prepurchase.domain.user.dto.LoginRequestDto;
-import org.example.prepurchase.domain.user.dto.SignupRequestDto;
-import org.example.prepurchase.domain.user.dto.UpdateImformationRequestDto;
-import org.example.prepurchase.domain.user.exception.DuplicateException;
-import org.example.prepurchase.global.auth.UserRoleEnum;
+import com.seonghun.module_user_service.domain.user.dao.UserRepository;
+import com.seonghun.module_user_service.domain.user.dto.LoginRequestDto;
+import com.seonghun.module_user_service.domain.user.dto.SignupRequestDto;
+import com.seonghun.module_user_service.domain.user.dto.UpdateImformationRequestDto;
+import com.seonghun.module_user_service.global.auth.UserRoleEnum;
+import com.seonghun.module_user_service.domain.user.domain.Users;
+import com.seonghun.module_user_service.domain.user.exception.DuplicateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
