+++ module_user_service/src/main/java/com/seonghun/module_user_service/domain/user/exception/DuplicateException.java
@@ -1,4 +1,4 @@
-package org.example.prepurchase.domain.user.exception;
+package com.seonghun.module_user_service.domain.user.exception;
 
 public class DuplicateException extends RuntimeException {
     public DuplicateException(String message) {
