+++ module_user_service/src/main/java/com/seonghun/module_user_service/global/error/ErrorDto.java
@@ -1,4 +1,4 @@
-package org.example.prepurchase.global.error;
+package com.seonghun.module_user_service.global.error;
 
 public class ErrorDto {
     private final String message;
